requirements.txt
google-generativeai>=0.8.5
openai>=1.0.0
rich>=14.0.0
python-dotenv>=1.0.0
prompt_toolkit>=3.0.0
pyperclip>=1.8.0